chore: handleAccessTokenExpired 함수 utils로 분리
